+++ src/binary-serializer/data-reader/zlib-reader.ts
@@ -4,14 +4,16 @@ import { DataReader } from "./interfaces";
 
 import { ArrayDataReader } from "./array-reader";
 
+import { toExactBuffer } from "../utils";
+
 export class ZlibDataReader extends ArrayDataReader {
   constructor(data: Uint8Array) {
     // ONI uses Ionic.Zlib.  More specifically, this:
     //  https://github.com/jstedfast/Ionic.Zlib/blob/master/Ionic.Zlib/ZlibStream.cs
 
     const deflated = inflate(data, {
-      windowBits: 15
+      windowBits: 15,
     });
-    super(deflated.buffer);
+    super(toExactBuffer(deflated));
   }
 }