+++ src/index.ts
@@ -20,7 +20,11 @@ export { tagReporter } from "./tagger";
 export {
   E_VERSION_MAJOR,
   E_VERSION_MINOR,
+  CURRENT_VERSION_MAJOR,
+  VERIFIED_VERSION_MINORS,
+  isVerifiedVersion,
 } from "./save-structure/version-validator";
+export type { VersionStrictness } from "./save-structure/version-validator";
 
 export interface ParseOptions extends SaveGameParserOptions {
   interceptor?: ParseInterceptor;