+++ src/save-structure/game-objects/game-object-behavior/known-behaviors/modifiers/parser.ts
@@ -119,7 +119,7 @@ function* unparseModifier<T extends ModificationInstance>(
   instance: T,
   modifierInstanceType: string,
   templateUnparser: TemplateUnparser
-) {
+): UnparseIterator {
   yield writeKleiString(instance.name);
 
   const token = yield writeDataLengthBegin();