+++ src/utils.ts
@@ -68,6 +68,6 @@ export function typed<T extends string>(s: T): T {
   return s;
 }
 
-export function typedKeys<T>(x: T): (keyof T)[] {
+export function typedKeys<T extends object>(x: T): (keyof T)[] {
   return Object.keys(x) as any;
 }